# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StockSense inventory management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: OTP request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /me:
    get:
      operationId: getMe
      tags: [auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /categories:
    get:
      operationId: listCategories
      tags: [products]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Product categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [products]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /warehouses:
    get:
      operationId: listWarehouses
      tags: [warehouses]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [warehouses]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "201":
          description: Warehouse created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
  /warehouses/{id}/locations:
    post:
      operationId: createLocation
      tags: [warehouses]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Location created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /stock:
    get:
      operationId: listStock
      tags: [inventory]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: productId
          in: query
          schema: { type: string }
        - name: locationId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Stock balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockBalance"
  /ledger:
    get:
      operationId: listLedger
      tags: [inventory]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: productId
          in: query
          schema: { type: string }
        - name: locationId
          in: query
          schema: { type: string }
        - name: referenceType
          in: query
          schema:
            $ref: "#/components/schemas/OperationType"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
      responses:
        "200":
          description: Stock ledger
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEntry"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: warehouseId
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/CategoryId"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OperationStatus"
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/OperationType"
      responses:
        "200":
          description: Dashboard KPIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /receipts:
    get:
      operationId: listReceipts
      tags: [operations]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receipt"
    post:
      operationId: createReceipt
      tags: [operations]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptInput"
      responses:
        "201":
          description: Receipt created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
  /receipts/{id}/validate:
    post:
      operationId: validateReceipt
      tags: [operations]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Receipt validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
...
  /transfers:
    get:
      operationId: listTransfers
      tags: [operations]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Internal transfers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transfer"
    post:
      operationId: createTransfer
      tags: [operations]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferInput"
      responses:
        "201":
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/validate:
    post:
      operationId: validateTransfer
      tags: [operations]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Transfer validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
        devOtp: { type: string }
...
    UserResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/User" }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
    SignUpInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordInput:
      type: object
      required: [email, otp, newPassword]
      properties:
        email: { type: string, format: email }
        otp: { type: string, minLength: 6, maxLength: 6 }
        newPassword: { type: string, minLength: 8 }
    Category:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
...